chore(program): clean program file
